Catch errors in Publisher
This also abstracts the tryEvent and tryEnd functions used in
create().
