+++ lib/source/Publisher.js
@@ -3,6 +3,10 @@
 /** @author John Hann */
 
 var Stream = require('../Stream');
+var sink = require('../sink/sink');
+
+var tryEvent = sink.tryEvent;
+var tryEnd   = sink.tryEnd;
 
 module.exports = Publisher;
 
@@ -13,39 +17,45 @@ function Publisher() {
 }
 
 Publisher.prototype.event = function(x) {
-	if(!this.active) {
-		return;
-	}
+	this._checkState();
+
 	var s = this.sinks.slice();
-	for(var i= 0, rec; i<s.length; ++i) {
+
+	for(var i=0, rec; i<s.length; ++i) {
 		rec = s[i];
-		rec.sink.event(rec.scheduler.now(), x);
+		tryEvent(rec.scheduler.now(), x, rec.sink);
 	}
 };
 
 Publisher.prototype.end = function(x) {
-	if(!this.active) {
-		return;
-	}
+	this._checkState();
 	this.active = false;
 
 	var s = this.sinks.slice();
-	for(var i= 0, rec; i<s.length; ++i) {
+	this.sinks = void 0;
+
+	for(var i=0, rec; i<s.length; ++i) {
 		rec = s[i];
-		rec.sink.end(rec.scheduler.now(), x);
+		tryEnd(rec.scheduler.now(), x, rec.sink);
 	}
 };
 
 Publisher.prototype.error = function(e) {
-	if(!this.active) {
-		return;
-	}
+	this._checkState();
 	this.active = false;
 
 	var s = this.sinks.slice();
-	for(var i= 0, rec; i<s.length; ++i) {
+	this.sinks = void 0;
+
+	for(var i=0, rec; i<s.length; ++i) {
 		rec = s[i];
 		rec.sink.error(rec.scheduler.now(), e);
+	}
+};
+
+Publisher.prototype._checkState = function() {
+	if(!this.active) {
+		throw new Error('stream ended');
 	}
 };
 
@@ -66,4 +76,5 @@ function PublisherDisposable(target, sinks) {
 
 PublisherDisposable.prototype.dispose = function() {
 	this.sinks.splice(this.sinks.indexOf(this.target), 1);
+	this.sinks = void 0;
 };
