+++ lib/source/create.js
@@ -7,6 +7,10 @@ var Disposable = require('../disposable/Disposable');
 var MulticastSource = require('./MulticastSource');
 var AsyncSource = require('./AsyncSource');
 var noop = require('../base').noop;
+var sink = require('../sink/sink');
+
+var tryEvent = sink.tryEvent;
+var tryEnd   = sink.tryEnd;
 
 exports.create = create;
 
@@ -35,19 +39,3 @@ SubscriberSource.prototype.run = function(sink, scheduler) {
 		sink.error(scheduler.now(), e);
 	}
 };
-
-function tryEvent(t, x, sink) {
-	try {
-		sink.event(t, x);
-	} catch(e) {
-		sink.error(t, e);
-	}
-}
-
-function tryEnd(t, x, sink) {
-	try {
-		sink.end(t, x);
-	} catch(e) {
-		sink.error(t, e);
-	}
-}
